Develop

+++ src/main/java/com/project/phonecaseshop/repository/MemberRepository.java
@@ -1,11 +1,13 @@
 package com.project.phonecaseshop.repository;
 
 import com.project.phonecaseshop.entity.Member;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Integer> {
     Member findByMemberEmail(String username);
     Member findByMemberNickname(String username);
 }
