+++ src/main/java/com/project/phonecaseshop/repository/RefreshTokenRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
 
     Optional<RefreshToken> findByMemberId(int memberId);
     Optional<RefreshToken> findByRefreshToken(String refreshToken);
